Address flakes on host.
These tests share a process on host, so we need the serial
annotation on every test case. That was missing from
utf8_and_bytestrings_share_the_cache which relies on counting items
in the global cache.

Also adds missing reset_global_cache() calls to new test cases.
Without this, host-side tests can have confusing results when one
failing test compromises the invariants of another.

Also changes the internal API for constructing a RawRepr to
require exclusive access to the cache so that reference counts
are guaranteed consistent. This is not a functional change given
the current constructors but is a useful guardrail as we refactor.

Multiply: flyweights
Change-Id: Ic00fc0d670fd21b9387217e9dcc7e9343cb592b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203549
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>