Add json schema support.
This adds json schema support for the logging command. It is done by
converting the main output from a `LogEntry` to a command-used
`CommandOutput` struct. The log entry is converted into this output for
simplicity's sake. The alternatives proved to be quite cumbersome (see
comments in the change).

Bug: 392657374

Change-Id: Idc38495247034dacf431ba200e769e6cd07e851b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222014
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>